+++ packages/packages/core/editor-editing-panel/package.json
@@ -44,6 +44,7 @@
     "@elementor/editor-controls": "3.33.0",
     "@elementor/editor-documents": "3.33.0",
     "@elementor/editor-elements": "3.33.0",
+    "@elementor/editor-interactions": "3.33.0",
     "@elementor/editor-panels": "3.33.0",
     "@elementor/editor-props": "3.33.0",
     "@elementor/editor-responsive": "3.33.0",
