+++ packages/packages/core/editor-editing-panel/src/components/interactions-tab.tsx
@@ -1,13 +1,11 @@
 import * as React from 'react';
 import { useCallback, useMemo, useState } from 'react';
 import { useElementInteractions } from '@elementor/editor-elements';
+import { EmptyState, PredefinedInteractionsList, usePopupStateContext } from '@elementor/editor-interactions';
 import { SessionStorageProvider } from '@elementor/session';
 
 import { useElement } from '../contexts/element-context';
 import { InteractionsProvider, useInteractionsContext } from '../contexts/interaction-context';
-import { usePopupStateContext } from '../contexts/popup-state-contex';
-import { EmptyState } from '../interactions/components/empty-state';
-import { PredefinedInteractionsList } from '../interactions/components/interactions-list';
 import { SectionsList } from './sections-list';
 
 export const InteractionsTab = () => {
