+++ packages/packages/core/editor-interactions/src/__tests__/predefined-interactions-list.test.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { render, screen } from '@testing-library/react';
 
-import { PopupStateProvider } from '../../contexts/popup-state-contex';
 import { PredefinedInteractionsList } from '../components/interactions-list';
+import { PopupStateProvider } from '../contexts/popup-state-contex';
 import { getInteractionsConfig } from '../utils/get-interactions-config';
 
 jest.mock( '../utils/get-interactions-config' );
