+++ packages/packages/core/editor-interactions/src/components/interaction-details.tsx
