+++ packages/packages/core/editor-interactions/src/utils/format-interaction-label.ts
